requirements.txt
absl-py==1.0.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
bs4==0.0.1
cachetools==5.1.0
cassandra-driver==3.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cssselect==1.1.0
fake-useragent==0.1.11
Flask==2.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.3.3
geomet==0.2.1.post1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-search-results==2.4.1
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==3.3
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
libclang==14.0.1
lxml==4.8.0
Markdown==3.2.1
MarkupSafe==2.1.1
numpy==1.21.6
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parse==1.19.0
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.4.1
six==1.14.0
soupsieve
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.0
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
websockets==10.3
Werkzeug==2.1.2
wikipedia==1.4.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.0
